requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
datastore-entity==0.2.0
Deprecated==1.2.13
Flask==2.1.1
Flask-Login==0.6.0
Flask-RESTful==0.3.9
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-core==2.3.0
google-cloud-datastore==1.15.4
google-cloud-ndb==1.11.1
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymemcache==3.5.2
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
requests==2.27.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
